Applied PR feedback
